register endpoint
