+++ app/model/user.go
@@ -1,0 +1,11 @@
+package model
+
+import (
+    "gorm.io/gorm"
+)
+
+type User struct {
+    gorm.Model
+    Username string
+    Password string
+}
